+++ src/Veldrid.SceneGraph/Viewer/Renderer.cs
@@ -198,7 +198,7 @@ namespace Veldrid.SceneGraph.Viewer
                     
                     _commandList.SetVertexBuffer(0, element.VertexBuffer);
                     
-                    _commandList.SetIndexBuffer(element.IndexBuffer, IndexFormat.UInt16);
+                    _commandList.SetIndexBuffer(element.IndexBuffer, IndexFormat.UInt32);
                     
                     _commandList.SetGraphicsResourceSet(0, _resourceSet);
 
@@ -310,7 +310,7 @@ namespace Veldrid.SceneGraph.Viewer
                     if (boundIndexBuffer != renderGroupElement.IndexBuffer)
                     {
                         // Set index buffer
-                        _commandList.SetIndexBuffer(renderGroupElement.IndexBuffer, IndexFormat.UInt16);
+                        _commandList.SetIndexBuffer(renderGroupElement.IndexBuffer, IndexFormat.UInt32);
                         boundIndexBuffer = renderGroupElement.IndexBuffer;
                     }
